Move GitHubLoader to its own "loading" directory
This also removes usage of the Octokit instance in the Core class.
