+++ src/environment/api.ts
@@ -1,7 +1,7 @@
 import { Badger } from "../badge/api";
+import { GitHubLoader } from "../loading/api";
 import { CrossScriptMessenger } from "../messaging/api";
 import { Notifier } from "../notifications/api";
-import { GitHubLoader } from "../state/github-loader";
 import { Store } from "../storage/api";
 
 export interface Environment {
